requirements.txt
# local package
-e .

# external requirements
pre-commit==2.21.0
click
Sphinx
coverage
awscli==1.27.41
flake8
python-dotenv>=0.5.1
dvc
nbdev
black
numpy
torch
wget
matplotlib
wandb
pytest
